+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SerializableTransactionManager.java
@@ -46,6 +46,7 @@ import com.palantir.lock.LockService;
 import com.palantir.lock.impl.LegacyTimelockService;
 import com.palantir.lock.v2.LockToken;
 import com.palantir.lock.v2.TimelockService;
+import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampService;
 
 public class SerializableTransactionManager extends SnapshotTransactionManager {
@@ -185,6 +186,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
     public static TransactionManager create(MetricsManager metricsManager,
             KeyValueService keyValueService,
             TimelockService timelockService,
+            TimestampManagementService timestampManagementService,
             LockService lockService,
             TransactionService transactionService,
             Supplier<AtlasDbConstraintCheckingMode> constraintModeSupplier,
@@ -205,6 +207,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
         return create(metricsManager,
                 keyValueService,
                 timelockService,
+                timestampManagementService,
                 lockService,
                 transactionService,
                 constraintModeSupplier,
@@ -228,6 +231,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
     public static TransactionManager create(MetricsManager metricsManager,
             KeyValueService keyValueService,
             TimelockService timelockService,
+            TimestampManagementService timestampManagementService,
             LockService lockService,
             TransactionService transactionService,
             Supplier<AtlasDbConstraintCheckingMode> constraintModeSupplier,
@@ -249,6 +253,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
                 metricsManager,
                 keyValueService,
                 timelockService,
+                timestampManagementService,
                 lockService,
                 transactionService,
                 constraintModeSupplier,
@@ -276,6 +281,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
     public static SerializableTransactionManager createForTest(MetricsManager metricsManager,
             KeyValueService keyValueService,
             TimestampService timestampService,
+            TimestampManagementService timestampManagementService,
             LockClient lockClient,
             LockService lockService,
             TransactionService transactionService,
@@ -290,6 +296,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
                 metricsManager,
                 keyValueService,
                 new LegacyTimelockService(timestampService, lockService, lockClient),
+                timestampManagementService,
                 lockService,
                 transactionService,
                 constraintModeSupplier,
@@ -309,6 +316,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
     public SerializableTransactionManager(MetricsManager metricsManager,
             KeyValueService keyValueService,
             TimelockService timelockService,
+            TimestampManagementService timestampManagementService,
             LockService lockService,
             TransactionService transactionService,
             Supplier<AtlasDbConstraintCheckingMode> constraintModeSupplier,
@@ -327,6 +335,7 @@ public class SerializableTransactionManager extends SnapshotTransactionManager {
                 metricsManager,
                 keyValueService,
                 timelockService,
+                timestampManagementService,
                 lockService,
                 transactionService,
                 constraintModeSupplier,