+++ docs/source/release_notes/release-notes.rst
@@ -50,6 +50,13 @@ develop
     *    - Type
          - Change
 
+    *    - |new|
+         - ``TransactionManager``s now expose a ``TimestampManagementService``, allowing clients to fast-forward timestamps when necessary.
+           This functionality is intended for libraries that extend AtlasDB functionality; it is unlikely that users should
+           directly require the TimestampManagementService.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3524>`__)
+
+
     *    - |fixed|
          - Targeted sweep no longer chokes if a table in the queue no longer exists,
            and was deleted by a different host while this host was online and sweeping.